Crate evmc_sys

Structs

evmc_address
Big-endian 160-bit hash suitable for keeping an Ethereum address.

evmc_bytes32
The fixed size array of 32 bytes.

evmc_host_interface
The Host interface.

evmc_message
The message describing an EVM call, including a zero-depth calls from a transaction origin.

evmc_result
The EVM code execution result.

evmc_tx_context
The transaction and block data for execution.

evmc_vm
The VM instance.

Enums

_bindgen_ty_1

evmc_call_kind
The kind of call-like instruction.

evmc_capabilities
Possible capabilities of a VM.

evmc_flags
The flags for ::evmc_message.

evmc_revision
EVM revision.

evmc_set_option_result
Possible outcomes of evmc_set_option.

evmc_status_code
The execution status code.

evmc_storage_status
The effect of an attempt to modify a contract storage item.

Constants

EVMC_ABI_VERSION

Type Definitions

__int32_t

__int64_t

__uint8_t

__uint32_t

evmc_account_exists_fn
Check account existence callback function.

evmc_call_fn
Pointer to the callback function supporting EVM calls.

evmc_capabilities_flagset
Alias for unsigned integer representing a set of bit flags of EVMC capabilities.

evmc_copy_code_fn
Copy code callback function.

evmc_destroy_fn
Destroys the VM instance.

evmc_emit_log_fn
Log callback function.

evmc_execute_fn
Executes the given code using the input from the message.

evmc_get_balance_fn
Get balance callback function.

evmc_get_block_hash_fn
Get block hash callback function.

evmc_get_capabilities_fn
Return the supported capabilities of the VM instance.

evmc_get_code_hash_fn
Get code hash callback function.

evmc_get_code_size_fn
Get code size callback function.

evmc_get_storage_fn
Get storage callback function.

evmc_get_tx_context_fn
Get transaction context callback function.

evmc_host_context

evmc_release_result_fn
Releases resources assigned to an execution result.

evmc_selfdestruct_fn
Selfdestruct callback function.

evmc_set_option_fn
Configures the VM instance.

evmc_set_storage_fn
Set storage callback function.

evmc_uint256be
The alias for evmc_bytes32 to represent a big-endian 256-bit integer.