Type Definition evmc_sys::evmc_execute_fn

type evmc_execute_fn = Option<unsafe extern "C" fn(vm: *mut evmc_vm, host: *const evmc_host_interface, context: *mut evmc_host_context, rev: evmc_revision, msg: *const evmc_message, code: *const u8, code_size: usize) -> evmc_result>;

Executes the given code using the input from the message.

This function MAY be invoked multiple times for a single VM instance.

@param vm The VM instance. This argument MUST NOT be NULL. @param host The Host interface. This argument MUST NOT be NULL unless the @p vm has the ::EVMC_CAPABILITY_PRECOMPILES capability. @param context The opaque pointer to the Host execution context. This argument MAY be NULL. The VM MUST pass the same pointer to the methods of the @p host interface. The VM MUST NOT dereference the pointer. @param rev The requested EVM specification revision. @param msg The call parameters. See ::evmc_message. This argument MUST NOT be NULL. @param code The reference to the code to be executed. This argument MAY be NULL. @param code_size The length of the code. If @p code is NULL this argument MUST be 0. @return The execution result.